Recurring appointment Fixes, take 2.
SUMMARY OF ISSUES DEALT WITH AND ADDITIONS -
-Appointment Report -
--Duplicate appointments -
For all recurring appointment dates that have been edited (e.g. change of status), two representations of these dates will appear in the report.
--Missing appointments -
All dates corresponding to recurring appointments set with options such as "2nd Tuesday" or "Last Friday" and on a date that is prior to the "From" date of the report will be missing from the report.
--Non existant and missing appointments -
This effects same type of appointment as in prior issue and locations with a positive UTC offset. If an appointment is set on a weekday that is the 4th occurance of that weekday and that weekday is the last weekday that has a 5th occurance, and the "Last <weekday>" option is selected, then a non existant appointment date will appear on  the report (with same date that the recurring appointment was set), also the first date of the series (first (or only) last <weekday>) will be missing (Calendar does not have this problem).
--Disageement with Calendar -
The output of the code that produces the report disagrees with the Calendar for some dates related to recurring appointments of same type as prior issue, set on 29th 30th or 31st of a month. The Calendar is getting some of these dates wrong, and the report does not.

-Patient Summary -
--Added support for recurring appointments.
--Appearance of appointments in Patient Summary -
2015-08-06 (Thu)
10:00 am o ( ? ) CMT
Established Patient
Dr Provider
"CMT" (green font) appears when there is a comment present, mousing over displays it.